Create Transition Probability Matrix Generator
This code generates the transition probability matrices of hidden and observable states in two modes. In the first mode, it generates a transition probability randomly. However, in the second mode, the pre-constructed transition probability matrices in the paper are generated. At the end, the stationary distribution of these two matrices will be calculated.